error handling: disable fatal error
For now, if the calls to Tika fail, the whole process fails. This is not ideal, but it is better than the alternative of silently failing. We should add a configuration option to disable fatal errors, and then we can add a retry mechanism.
We let the user to decide if he wants to disable fatal errors or not,
using config TikaAllTheFiles_ThrowExceptionOnTikaFailure