fix: Make this actually work with MW 1.39(+) and PHP 8.2.
MediaHandler's public/protected interface has not really changed from 1.37
through 1.41, so getting 1.39+ to work only requires extending the list
of known releases.

PHP 8.2 is *not* happy with external access of private members/methods,
or static calls of non-static methods, and rightfully so!  This is fixed
by making some things protected, and changing to a non-static call.

*Only tested with MW 1.39 and PHP 8.2.*